Cleanup

+++ portal/assignment.py
@@ -2,77 +2,35 @@ from flask import Flask, render_template, Blueprint, request, redirect, url_for,
 
 from . import db
 
-from portal.auth import login_required, admin
+from portal.auth import login_required, admin, validate, validate_text, validate_date, validate_number
 from portal.teacher import bp
-@bp.route('/assignments/create', methods=('GET', 'POST'))
-def create_assignments():
-    if request.method == 'POST':
-        name = request.form['name']
-        description = request.form['description']
-        points = request.form['points']
-        course = request.form['course']
-        with db.get_db() as con:
-            with con.cursor() as cur:
-                cur.execute("""
-                    INSERT INTO assignments (name, description, points, course_id)
-                    VALUES (%s, %s, %s, %s)
-                """, (name, description, points, course))
-                print(g.user['id'])
-                return redirect(url_for('teacher.assignments'))
-    with db.get_db() as con:
-        with con.cursor() as cur:
-            cur.execute("""
-                SELECT * FROM courses
-                WHERE teacher_id = %s
-            """, (g.user['id'],))
-            courses = cur.fetchall()
-    return render_template('create-assignments.html', courses=courses)
-    
-@bp.route('/assignments/assign/submit', methods=('GET', 'POST'))
-def assign_submit():
-     if request.method == 'POST':
-        date = request.form['date']
-        assign_id = request.form['assign_id']
-        session_id = request.form['session_id']
-        with db.get_db() as con:
-            with con.cursor() as cur:
-                cur.execute("""
-                    INSERT INTO session_assignments (session_id, assignment_id, due_date)
-                    VALUES (%s, %s, %s)
-                    """, (session_id, assign_id, date, ))
-     return redirect(url_for('teacher.sessions'))
-@bp.route('/assignments/assign', methods=('GET', 'POST'))
-def assign_work():
-    if request.method == 'POST':
-        session_id = request.form['session_id']
-        with db.get_db() as con:
-            with con.cursor() as cur:
-                cur.execute("""
-                    SELECT * FROM assignments
-                    WHERE course_id IN (SELECT course_id FROM sessions WHERE id = %s)
-                    AND id NOT IN (SELECT assignment_id FROM session_assignments
-                                   WHERE session_id = %s)
-                """, (session_id, session_id))
-                assigns = cur.fetchall()
-        return render_template('assign-work.html', assigns=assigns, session_id=session_id)
-    return redirect(url_for('teacher.sessions'))
+from portal.session import bp
 
 @bp.route('/assignments', methods=('GET', 'POST'))
 @login_required
 @admin
 def assignments():
     if request.method == 'POST':
-        with db.get_db() as con:
-            with con.cursor() as cur:
-                for item in request.form.getlist('id'):
-                    cur.execute("""
-                        DELETE FROM session_assignments
-                        WHERE assignment_id = %s
-                    """, (item,))
+
+        items = []
+        error = None
+
+        for item in request.form.getlist('id'):
+            if validate(item, 'assignments'):
+                items.append(int(item))
+            else:
+                error = 'Something went wrong.'
+                break
+
+        if not error:
+            with db.get_db() as con:
+                with con.cursor() as cur:
                     cur.execute("""
                         DELETE FROM assignments
-                        WHERE id = %s
-                    """, (item,))
+                        WHERE id = ANY(%s)
+                    """, (items,))
+        else:
+            flash(error)
 
     with db.get_db() as con:
         with con.cursor() as cur:
@@ -81,10 +39,11 @@ def assignments():
                 FROM assignments a JOIN courses c
                 ON a.course_id = c.id
                 WHERE c.teacher_id = %s
+                ORDER BY a.id
             """, (g.user['id'],))
             assignments = cur.fetchall()
 
-    return render_template('assignments.html', assignments=assignments)
+    return render_template('layouts/teacher/assignments/assignments.html', assignments=assignments)
 
 @bp.route('/assignments/edit', methods=('GET', 'POST'))
 @login_required
@@ -92,17 +51,21 @@ def assignments():
 def edit_assignments():
     if request.method == 'POST':
         assignment_id = request.form['edit']
-        with db.get_db() as con:
-            with con.cursor() as cur:
-                cur.execute("""
-                    SELECT * FROM assignments
-                    WHERE id = %s
-                    """, (assignment_id,))
-                info = cur.fetchone()
-        return render_template('assignments/edit-assignments.html', info=info)
+        if validate(assignment_id, 'assignments'):
+            with db.get_db() as con:
+                with con.cursor() as cur:
+                    cur.execute("""
+                        SELECT * FROM assignments
+                        WHERE id = %s
+                        """, (assignment_id,))
+                    info = cur.fetchone()
+            return render_template('layouts/teacher/assignments/edit-assignments.html', info=info)
+        else:
+            flash('Something went wrong')
+
     return redirect(url_for('teacher.assignments'))
 
-@bp.route('/assignments/submit', methods=('GET', 'POST'))
+@bp.route('/assignments/edit/submit', methods=('GET', 'POST'))
 @login_required
 @admin
 def submit_assignments():
@@ -111,14 +74,106 @@ def submit_assignments():
         desc = request.form['description']
         points = request.form['points']
         id = request.form['submit']
-        with db.get_db() as con:
-            with con.cursor() as cur:
-                cur.execute("""
-                    UPDATE assignments
-                    SET name = %s, description = %s, points = %s
-                    WHERE id = %s
-                """, (name, desc, points, id))
+        if (
+            validate(id, 'assignments') and
+            validate_text(name, 50) and
+            validate_text(desc, 300) and
+            validate_number(points, 100000)
+            ):
+            with db.get_db() as con:
+                with con.cursor() as cur:
+                    cur.execute("""
+                        UPDATE assignments
+                        SET name = %s, description = %s, points = %s
+                        WHERE id = %s
+                    """, (name, desc, points, id))
+        else:
+            flash('Something went wrong.')
+
     return redirect(url_for('teacher.assignments'))
+
+@bp.route('/assignments/create', methods=('GET', 'POST'))
+@login_required
+@admin
+def create_assignments():
+    if request.method == 'POST':
+        name = request.form['name']
+        desc = request.form['description']
+        points = request.form['points']
+        course = request.form['course']
+
+        if (
+            validate_text(name, 50) and
+            validate_text(desc, 300) and
+            validate_number(points, 100000) and
+            validate(course, 'courses')
+            ):
+            with db.get_db() as con:
+                with con.cursor() as cur:
+                    cur.execute("""
+                        INSERT INTO assignments (name, description, points, course_id)
+                        VALUES (%s, %s, %s, %s)
+                    """, (name, desc, points, course))
+
+                    return redirect(url_for('teacher.assignments'))
+        else:
+            flash("Something went wrong.")
+
+    with db.get_db() as con:
+        with con.cursor() as cur:
+            cur.execute("""
+                SELECT * FROM courses
+                WHERE teacher_id = %s
+            """, (g.user['id'],))
+            courses = cur.fetchall()
+
+    return render_template('layouts/teacher/assignments/create-assignments.html', courses=courses)
+
+@bp.route('/assignments/assign', methods=('GET', 'POST'))
+@login_required
+@admin
+def assign_work():
+    if request.method == 'POST':
+        session_id = request.form['session_id']
+        if validate(session_id, 'sessions'):
+            with db.get_db() as con:
+                with con.cursor() as cur:
+                    cur.execute("""
+                        SELECT * FROM assignments
+                        WHERE course_id IN (SELECT course_id FROM sessions WHERE id = %s)
+                        AND id NOT IN (SELECT assignment_id FROM session_assignments
+                                       WHERE session_id = %s)
+                    """, (session_id, session_id))
+                    assigns = cur.fetchall()
+            return render_template('layouts/teacher/assignments/assign-work.html', assigns=assigns, session_id=session_id)
+        else:
+            flash('Something went wrong.')
+
+    return redirect(url_for('teacher.sessions'))
+
+@bp.route('/assignments/assign/submit', methods=('GET', 'POST'))
+@login_required
+@admin
+def assign_submit():
+    if request.method == 'POST':
+        date = request.form['date']
+        assign_id = request.form['assign_id']
+        session_id = request.form['session_id']
+
+        if (
+            validate(assign_id, 'assignments') and
+            validate(session_id, 'sessions') and
+            validate_date(date)
+            ):
+            with db.get_db() as con:
+                with con.cursor() as cur:
+                    cur.execute("""
+                        INSERT INTO session_assignments (session_id, assignment_id, due_date)
+                        VALUES (%s, %s, %s)
+                    """, (session_id, assign_id, date))
+        else:
+            flash('Something went wrong.')
+    return redirect(url_for('teacher.sessions'))
 
 @bp.route('/assignments/grade', methods=('GET','POST'))
 @login_required
@@ -126,21 +181,28 @@ def submit_assignments():
 def grade():
     if request.method == 'POST':
         code = request.form['grade']
-        with db.get_db() as con:
-            with con.cursor() as cur:
-                cur.execute("""
-                SELECT r.name, r.description, r.points, u.first_name, u.last_name, u.id, a.work_id
-                FROM session_assignments a JOIN assignments r
-                ON a.assignment_id = r.id
-                JOIN roster d
-                ON a.session_id = d.session_id
-                JOIN users u
-                ON d.student_id = u.id
-                WHERE a.assignment_id = %s
-                """, (code, ))
-                informations = cur.fetchall()
-        return render_template('assignments/teacher-assignments.html', informations=informations)
+
+        if validate(code, 'assignments'):
+            with db.get_db() as con:
+                with con.cursor() as cur:
+                    cur.execute("""
+                        SELECT r.name, r.description, r.points, u.first_name, u.last_name, u.id, a.work_id
+                        FROM session_assignments a JOIN assignments r
+                        ON a.assignment_id = r.id
+                        JOIN roster d
+                        ON a.session_id = d.session_id
+                        JOIN users u
+                        ON d.student_id = u.id
+                        WHERE a.assignment_id = %s
+                    """, (code,))
+                    informations = cur.fetchall()
+            return render_template('layouts/teacher/assignments/teacher-assignments.html', informations=informations)
+
+        else:
+            flash('Something went wrong.')
+
     return redirect(url_for('teacher.courses'))
+
 
 @bp.route('/assignments/view', methods=('GET', 'POST'))
 @login_required
@@ -148,55 +210,58 @@ def grade():
 def view_assignments():
     if request.method == 'POST':
         code = request.form['view-grade']
-        with db.get_db() as con:
-            with con.cursor() as cur:
-                cur.execute("""
-                SELECT a.name, a.description, a.points, c.course_name, a.id, p.work_id
-                FROM sessions s JOIN session_assignments p
-                ON s.id = p.session_id
-                JOIN assignments a
-                ON p.assignment_id = a.id
-                JOIN courses c
-                ON c.id = s.course_id
-                WHERE s.id = %s
-                """, (code,))
-                assignments = cur.fetchall()
-        return render_template('assignments/view-assignments.html', assignments=assignments)
-
+        if validate(code, 'sessions'):
+            with db.get_db() as con:
+                with con.cursor() as cur:
+                    cur.execute("""
+                        SELECT a.name, a.description, a.points, c.course_name, a.id, p.work_id
+                        FROM sessions s JOIN session_assignments p
+                        ON s.id = p.session_id
+                        JOIN assignments a
+                        ON p.assignment_id = a.id
+                        JOIN courses c
+                        ON c.id = s.course_id
+                        WHERE s.id = %s
+                    """, (code,))
+                    assignments = cur.fetchall()
+            return render_template('layouts/teacher/assignments/view-assignments.html', assignments=assignments)
+        else:
+            flash("Something went wrong.")
     return redirect(url_for('teacher.courses'))
 
-@bp.route('/grade/submission', methods=('GET', 'POST'))
+@bp.route('assignments/grade/submission', methods=('GET', 'POST'))
 @login_required
 @admin
 def grade_submission():
     if request.method == 'POST':
-        grade = request.form['grade-submission']
-        user = list(request.form['submission'])
+        grade = request.form['grade']
+        student_id = request.form['submission']
+        assignment_id = request.form['assignment_id']
         with db.get_db() as con:
-                with con.cursor() as cur:
-                    cur.execute("""
-                    SELECT * FROM assignment_grades
-                    WHERE owner_id = %s AND assigned_id = %s
-                    """, (user[1], user[4], ))
-                    search = cur.fetchall()
+            with con.cursor() as cur:
+                cur.execute("""
+                SELECT * FROM assignment_grades
+                WHERE owner_id = %s AND assigned_id = %s
+                """, (student_id, assignment_id))
+                search = cur.fetchall()
         if not search:
             with db.get_db() as con:
                 with con.cursor() as cur:
                     cur.execute("""
                     INSERT INTO assignment_grades(owner_id, assigned_id, grades) VALUES (%s, %s, %s);
                     SELECT * FROM assignment_grades
-                    """, (user[1], user[4], grade, ))
+                    """, (student_id, assignment_id, grade))
                     res = cur.fetchall()
                     return redirect(url_for('teacher.sessions'))
         else :
               with db.get_db() as con:
                   with con.cursor() as cur:
                       cur.execute("""
-                      UPDATE assignment_grades
-                      SET grades = %s
-                      WHERE owner_id = %s AND assigned_id = %s;
-                      SELECT * FROM assignment_grades
-                      """, (grade ,user[1], user[4],))
+                          UPDATE assignment_grades
+                          SET grades = %s
+                          WHERE owner_id = %s AND assigned_id = %s;
+                          SELECT * FROM assignment_grades
+                      """, (grade ,student_id, assignment_id))
                       res = cur.fetchall()
                       return redirect(url_for('teacher.sessions'))
     return redirect(url_for('teacher.courses'))
@@ -225,5 +290,5 @@ def assignment_grades():
                 """, (assignment_id,))
                 assignment_name = cur.fetchone()
 
-        return render_template('assignment-grades.html', assignment=assignment, assignment_name=assignment_name)
+        return render_template('layouts/teacher/assignments/assignment-grades.html', assignment=assignment, assignment_name=assignment_name)
     return redirect(url_for('teacher.home'))
